Gitignore modified to add license and remove *sql, renamed package name
