+++ .gitignore
@@ -1,5 +1,7 @@
 # Temp Folders #
 tmp/
+target/
+License
 
 # Compiled source #
 ###################
@@ -26,7 +28,6 @@ tmp/
 # Logs and databases #
 ######################
 *.log
-*.sql
 *.sqlite
 
 # OS generated files #
